five category

+++ app/src/main/java/rpr/events/Pentarnery.java
@@ -44,13 +44,15 @@ public class Pentarnery extends Fragment {
     private static final String TAG_TIME = "time";
     private static final String TAG_VENUE = "venue";
     private static final String TAG_DETAILS = "details";
+    private static String usertype_id;
+
+    UserSessionManager session;
 
     private static ProgressDialog pDialog;
     private ListView lv;
 
     // URL to get events JSON
-    private static String ListURL = "http://10.1.1.19/~2015csb1021/php/ListAll.php";
-
+    private static String ListURL = "http://10.1.1.19/~2015csb1021/event/listAll.php";
     ArrayList<HashMap<String, String>> eventList;
 
     @Override
@@ -60,11 +62,14 @@ public class Pentarnery extends Fragment {
         //change R.layout.yourlayoutfilename for each of your fragments
         context = getActivity();
         createDatabase();
+        session = new UserSessionManager(getContext());
+        HashMap<String, String> user = session.getUserDetails();
+        usertype_id = user.get(UserSessionManager.KEY_USERTYPE_ID);
         return inflater.inflate(R.layout.pentarnery_layout, container, false);
 
     }
     protected void createDatabase(){
-        db=getActivity().openOrCreateDatabase("EventDB", Context.MODE_PRIVATE, null);
+        db=getActivity().openOrCreateDatabase("EventDB4", Context.MODE_PRIVATE, null);
         db.execSQL("CREATE TABLE IF NOT EXISTS Events(event_id INTEGER NOT NULL, name VARCHAR NOT NULL,venue VARCHAR NOT NULL, time TIMESTAMP NOT NULL, details VARCHAR NOT NULL  );");
     }
 
@@ -77,14 +82,6 @@ public class Pentarnery extends Fragment {
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        //you can set the title for your toolbar here for different fragments different titles
-        getActivity().setTitle("Menu 1");
-//        ProgressDialog pDialog;
-//        ListView lv;
-
-        // URL to get events JSON
-        //String ListURL = "http://10.1.1.19/~2015csb1021/php/ListAll.php";
-
         eventList = new ArrayList<>();
         lv = (ListView) getView().findViewById(R.id.listView);
 
@@ -111,7 +108,7 @@ public class Pentarnery extends Fragment {
             HttpHandler sh = new HttpHandler();
 
             // Making a request to url and getting response
-            String jsonStr = sh.makeServiceCall(ListURL);
+            String jsonStr = sh.makeServiceCall(ListURL+"?usertype_id="+"?usertype_id="+usertype_id+"&category_id=4");
 
             Log.e(TAG, "Response from url: " + jsonStr);
 
@@ -119,47 +116,50 @@ public class Pentarnery extends Fragment {
                 try {
                     JSONObject jsonObj = new JSONObject(jsonStr);
 
+                    if (jsonObj.getBoolean("success") == true) {
+                        JSONArray events = jsonObj.getJSONArray("events");
 
-                    JSONArray events = jsonObj.getJSONArray("events");
 
+                        db.execSQL("DELETE FROM Events");
+                        for (int i = 0; i < events.length(); i++) {
+                            JSONObject c = events.getJSONObject(i);
 
-                    db.execSQL("DELETE FROM Events");
-                    for (int i = 0; i < events.length(); i++) {
-                        JSONObject c = events.getJSONObject(i);
+                            String event_id = c.getString(TAG_EVENT_ID);
+                            String name = c.getString(TAG_NAME);
+                            String time = c.getString(TAG_TIME);
+                            String venue = c.getString(TAG_VENUE);
+                            String details = c.getString(TAG_DETAILS);
 
-                        String event_id = c.getString(TAG_EVENT_ID);
-                        String name = c.getString(TAG_NAME);
-                        String time = c.getString(TAG_TIME);
-                        String venue = c.getString(TAG_VENUE);
-                        String details = c.getString(TAG_DETAILS);
+                            insertIntoDB(event_id, name, time, venue, details);
 
-                        insertIntoDB(event_id, name, time, venue, details);
+                            // tmp hash map for single event
+                            HashMap<String, String> event = new HashMap<>();
 
-                        // tmp hash map for single event
-                        HashMap<String, String> event = new HashMap<>();
+                            // adding each child node to HashMap key => value
+                            event.put(TAG_EVENT_ID, event_id);
+                            event.put(TAG_NAME, name);
+                            event.put(TAG_TIME, time);
+                            event.put(TAG_VENUE, venue);
+                            event.put(TAG_DETAILS, details);
 
-                        // adding each child node to HashMap key => value
-                        event.put(TAG_EVENT_ID, event_id);
-                        event.put(TAG_NAME, name);
-                        event.put(TAG_TIME, time);
-                        event.put(TAG_VENUE, venue);
-                        event.put(TAG_DETAILS, details);
-
-                        // adding event to event list
-                        eventList.add(event);
+                            // adding event to event list
+                            eventList.add(event);
+                        }
+                        db.close();
                     }
-                    db.close();
+                    else{
+                        Toast.makeText(context.getApplicationContext(),
+                                "No events found",
+                                Toast.LENGTH_LONG)
+                                .show();
+                    }
 
                 } catch (final JSONException e) {
-                    Log.e(TAG, "Json parsing error: " + e.getMessage());
 
                     getActivity().runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            Toast.makeText(context.getApplicationContext(),
-                                    "Json parsing error: " + e.getMessage(),
-                                    Toast.LENGTH_LONG)
-                                    .show();
+
                         }
                     });
